apps/api/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
coverage==7.6.9
welleng==0.8.5
numpy>=1.24.0
openpyxl==3.1.5
